pyproject.toml
[project]
name = "z-ml-ops-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bitsandbytes>=0.49.1",
    "cloudpickle>=3.1.2",
    "datasets>=4.5.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.5",
    "google-genai>=1.63.0",
    "hyperopt>=0.2.7",
    "ipykernel>=7.2.0",
    "keras>=3.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "mlflow>=3.9.0",
    "numpy>=2.4.2",
    "openai>=2.20.0",
    "pandas>=2.3.3",
    "peft>=0.18.1",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "ragas>=0.4.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "sentence-transformers>=5.2.2",
    "tensorflow>=2.20.0",
    "torch>=2.10.0",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
]
